Factor out "fake" numbers for tests
...into their own reusable header file.
Move conj_if_needed, real_part, and imag_part tests
out of proxy_refs.cpp into a new test file.
