+++ tests/native/CMakeLists.txt
@@ -33,6 +33,7 @@ linalg_add_test(scale)
 linalg_add_test(scaled)
 linalg_add_test(swap)
 linalg_add_test(symm)
+linalg_add_test(test_test_numbers)
 linalg_add_test(transposed)
 linalg_add_test(trmm)
 linalg_add_test(trsm)
