+++ tests/native/proxy_refs.cpp
@@ -1,181 +1,9 @@
-// To try using subscript operator comment in macro below
-// the header will by default also check for the feature macro, and enable it
-// defining the macro to 0 will overwrite the automatic setting
-// x86-64 clang (experimental auto NSDMI) supports the operator, but you need
-// to explicitly comment in below macro
-//#define MDSPAN_USE_BRACKET_OPERATOR 1
-
-// To force enable operator() comment in the macro below
-// You can enable both at the same time. 
-//#define MDSPAN_USE_PAREN_OPERATOR 0
-
+#define MDSPAN_USE_PAREN_OPERATOR 1
 #define P1673_CONJUGATED_SCALAR_ARITHMETIC_OPERATORS_REFERENCE_OVERLOADS 1
 
 #include "gtest/gtest.h"
+#include "./test_numbers.hpp"
 #include <experimental/linalg>
-
-///////////////////////////////////////////////////////////
-// Custom real number type for tests
-///////////////////////////////////////////////////////////
-
-struct FakeRealNumber {
-    float value;
-
-#ifdef __cpp_impl_three_way_comparison
-    bool operator==(const FakeRealNumber&) const = default;
-#else
-  friend bool operator==(const FakeRealNumber& x, const FakeRealNumber& y)
-  {
-    return x.value == y.value;
-  }
-#endif
-};
-// Non-arithmetic types need a definition of conj,
-// else conjugated_scalar won't compile with them.
-FakeRealNumber conj(const FakeRealNumber& x) { return x; }
-
-// Custom complex number type
-
-struct FakeComplex {
-    double real;
-    double imag;
-
-#ifdef __cpp_impl_three_way_comparison
-    bool operator==(const FakeComplex&) const = default;
-#else
-  friend bool operator==(const FakeComplex& x, const FakeComplex& y)
-  {
-    return x.real == y.real && x.imag == y.imag;
-  }
-#endif
-
-    constexpr FakeComplex& operator+=(const FakeComplex& other)
-    {
-        real += other.real;
-        imag += other.imag;
-        return *this;
-    }
-    constexpr FakeComplex& operator-=(const FakeComplex& other)
-    {
-        real -= other.real;
-        imag -= other.imag;
-        return *this;
-    }
-    constexpr FakeComplex& operator*=(const FakeComplex& other)
-    {
-        real = real * other.real - imag * other.imag;
-        imag = imag * other.real + real * other.imag;
-        return *this;
-    }
-    constexpr FakeComplex& operator/=(const FakeComplex& other)
-    {
-        // just for illustration; please don't implement it this way.
-        const auto other_mag = other.real * other.real + other.imag * other.imag;
-        real = (real * other.real + imag * other.imag) / other_mag;
-        imag = (imag * other.real - real * other.imag) / other_mag;
-        return *this;
-    }
-
-    constexpr FakeComplex& operator+=(const double other)
-    {
-        real += other;
-        return *this;
-    }
-    constexpr FakeComplex& operator-=(const double other)
-    {
-        real -= other;
-        return *this;
-    }
-    constexpr FakeComplex& operator*=(const double other)
-    {
-        real *= other;
-        imag *= other;
-        return *this;
-    }
-    constexpr FakeComplex& operator/=(const double other)
-    {
-        real /= other;
-        imag /= other;
-        return *this;
-    }
-};
-
-// Unary operators
-
-FakeComplex operator+( const FakeComplex& val )
-{
-    return val;
-}
-FakeComplex operator-( const FakeComplex& val )
-{
-    return {-val.real, -val.imag};
-}
-
-// Binary homogeneous operators
-
-FakeComplex operator+(const FakeComplex& z, const FakeComplex& w)
-{
-    return {z.real + w.real, z.imag + w.imag};
-}
-FakeComplex operator-(const FakeComplex& z, const FakeComplex& w)
-{
-    return {z.real - w.real, z.imag - w.imag};
-}
-FakeComplex operator*(const FakeComplex& z, const FakeComplex& w)
-{
-    return {z.real * w.real - z.imag * w.imag,
-        z.imag * w.real + z.real * w.imag};
-}
-FakeComplex operator/(const FakeComplex& z, const FakeComplex& w)
-{
-    // just for illustration; please don't implement it this way.
-    const auto w_mag = w.real * w.real + w.imag * w.imag;
-    return {(z.real * w.real + z.imag * w.imag) / w_mag,
-        (z.imag * w.real - z.real * w.imag) / w_mag};
-}
-
-// Binary (complex,real) operators
-
-FakeComplex operator+(const FakeComplex& z, const double w)
-{
-    return {z.real + w, z.imag};
-}
-FakeComplex operator-(const FakeComplex& z, const double w)
-{
-    return {z.real - w, z.imag};
-}
-FakeComplex operator*(const FakeComplex& z, const double w)
-{
-    return {z.real * w, z.imag * w};
-}
-FakeComplex operator/(const FakeComplex& z, const double w)
-{
-    return {z.real / w, z.imag / w};
-}
-
-// Binary (real,complex) operators
-
-FakeComplex operator+(const double z, const FakeComplex& w)
-{
-    return {z + w.real, z + w.imag};
-}
-FakeComplex operator-(const double z, const FakeComplex& w)
-{
-    return {z - w.real, -w.imag};
-}
-FakeComplex operator*(const double z, const FakeComplex& w)
-{
-    return {z * w.real, z * w.imag};
-}
-FakeComplex operator/(const double z, const FakeComplex& w)
-{
-    // just for illustration; please don't implement it this way.
-    const auto w_mag = w.real * w.real + w.imag * w.imag;
-    return {
-        (z * w.real) / w_mag,
-        (-(z * w.imag)) / w_mag
-    };
-}
 
 // Specialize test helper traits (P1673 does NOT need these)
 namespace test_helpers {
@@ -204,18 +32,6 @@ static constexpr bool is_atomic_ref_not_arithmetic_v<std::atomic_ref<U>> = ! std
 #endif
 
 } // namespace test_helpers
-
-// FakeComplex conj implementation
-FakeComplex conj(const FakeComplex& z) { return {z.real, -z.imag}; }
-
-// FakeComplex abs implementation
-auto abs(const FakeComplex& z) { return sqrt(z.real * z.real + z.imag * z.imag); }
-
-// FakeComplex real implementation
-auto real(const FakeComplex& z) { return z.real; }
-
-// FakeComplex imag implementation
-auto imag(const FakeComplex& z) { return z.imag; }
 
 ///////////////////////////////////////////////////////////
 // conj_if_needed tests
@@ -1078,148 +894,6 @@ void test_FakeComplex_scaled_scalar(const ScalingFactor& scalingFactor)
 }
 
 namespace {
-  TEST(proxy_refs, conj_if_needed)
-  {
-    test_complex_conj_if_needed<float>();
-    test_complex_conj_if_needed<double>();
-    test_complex_conj_if_needed<long double>();
-
-    test_real_conj_if_needed<float>();
-    test_real_conj_if_needed<double>();
-    test_real_conj_if_needed<long double>();
-
-    test_real_conj_if_needed<int32_t>();
-    test_real_conj_if_needed<uint32_t>();
-    test_real_conj_if_needed<int64_t>();
-    test_real_conj_if_needed<uint64_t>();
-  }
-
-  template<class R>
-  void test_imag_part_complex()
-  {
-    using std::experimental::linalg::impl::imag_part;
-    std::complex<R> z{R(3.0), R(4.0)};
-    auto z_imag = imag_part(z);
-    EXPECT_EQ(z_imag, R(4.0));
-    static_assert(std::is_same_v<decltype(z_imag), R>);
-  }
-  template<class T>
-  void test_imag_part_floating_point()
-  {
-    using std::experimental::linalg::impl::imag_part;
-    T x = 9.0;
-    auto x_imag = imag_part(x);
-    EXPECT_EQ(x_imag, T(0.0));
-    static_assert(std::is_same_v<decltype(x_imag), T>);
-  }
-  template<class T>
-  void test_imag_part_integral()
-  {
-    using std::experimental::linalg::impl::imag_part;
-    T x = 3;
-    auto x_imag = imag_part(x);
-    EXPECT_EQ(x_imag, T(0));
-    static_assert(std::is_same_v<decltype(x_imag), T>);
-  }
-
-  TEST(proxy_refs, imag_part)
-  {
-    test_imag_part_complex<float>();
-    test_imag_part_complex<double>();
-    test_imag_part_complex<long double>();
-    
-    test_imag_part_floating_point<float>();
-    test_imag_part_floating_point<double>();
-    test_imag_part_floating_point<long double>();    
-
-    test_imag_part_integral<int8_t>();
-    test_imag_part_integral<uint8_t>();    
-    test_imag_part_integral<int16_t>();
-    test_imag_part_integral<uint16_t>();    
-    test_imag_part_integral<int32_t>();
-    test_imag_part_integral<uint32_t>();    
-    test_imag_part_integral<int64_t>();
-    test_imag_part_integral<uint64_t>();    
-
-    {
-      using std::experimental::linalg::impl::imag_part;
-      FakeComplex z{3.0, 4.0};
-      auto z_imag = imag_part(z);
-      EXPECT_EQ(z_imag, 4.0);
-      static_assert(std::is_same_v<decltype(z_imag), decltype(z.imag)>);
-    }
-    {
-      using std::experimental::linalg::impl::imag_part;
-      FakeRealNumber x{3.0};
-      auto x_imag = imag_part(x);
-      EXPECT_EQ(x_imag, FakeRealNumber{});
-      static_assert(std::is_same_v<decltype(x_imag), FakeRealNumber>);
-    }
-  }
-
-  template<class R>
-  void test_real_part_complex()
-  {
-    using std::experimental::linalg::impl::real_part;
-    std::complex<R> z{R(3.0), R(4.0)};
-    auto z_imag = real_part(z);
-    EXPECT_EQ(z_imag, R(3.0));
-    static_assert(std::is_same_v<decltype(z_imag), R>);
-  }
-  template<class T>
-  void test_real_part_floating_point()
-  {
-    using std::experimental::linalg::impl::real_part;
-    T x = 9.0;
-    auto x_imag = real_part(x);
-    EXPECT_EQ(x_imag, T(9.0));
-    static_assert(std::is_same_v<decltype(x_imag), T>);
-  }
-  template<class T>
-  void test_real_part_integral()
-  {
-    using std::experimental::linalg::impl::real_part;
-    T x = 3;
-    auto x_imag = real_part(x);
-    EXPECT_EQ(x_imag, T(3));
-    static_assert(std::is_same_v<decltype(x_imag), T>);
-  }
-
-  TEST(proxy_refs, real_part)
-  {
-    test_real_part_complex<float>();
-    test_real_part_complex<double>();
-    test_real_part_complex<long double>();
-    
-    test_real_part_floating_point<float>();
-    test_real_part_floating_point<double>();
-    test_real_part_floating_point<long double>();    
-
-    test_real_part_integral<int8_t>();
-    test_real_part_integral<uint8_t>();    
-    test_real_part_integral<int16_t>();
-    test_real_part_integral<uint16_t>();    
-    test_real_part_integral<int32_t>();
-    test_real_part_integral<uint32_t>();    
-    test_real_part_integral<int64_t>();
-    test_real_part_integral<uint64_t>();    
-
-    {
-      using std::experimental::linalg::impl::real_part;
-      FakeComplex z{3.0, 4.0};
-      auto z_imag = real_part(z);
-      EXPECT_EQ(z_imag, 3.0);
-      static_assert(std::is_same_v<decltype(z_imag), decltype(z.imag)>);
-    }
-    {
-      using std::experimental::linalg::impl::real_part;
-      FakeRealNumber x{3.0};
-      auto x_real = real_part(x);
-      EXPECT_EQ(x_real, FakeRealNumber{3.0});
-      static_assert(std::is_same_v<decltype(x_real), FakeRealNumber>);
-    }
-  }
-  
   TEST(proxy_refs, conjugated_scalar)
   {
     test_complex_conjugated_scalar<float>();
